ssh: allow linking against libssh2 as well as detecting at runtime
There is no end to the variation in the way people decide to package their
applications, and there seems to be no larger variety than in the way they
decide to deal with libssh2.

Re-introduce looking for the library at compile time, while adding the
USE_SSH_RUNTIME flag to enable detection at runtime.